requirements.txt
djangorestframework
mysqlclient
ipython
Pillow